add: eth registory
