+++ packages/issuer/README.md
@@ -18,10 +18,6 @@ https://community.blockcerts.org/t/introducing-smart-contracts-to-blockcerts/236
 試した環境はubuntu 20.04, python 3.9.13です. pyenvを使っている場合は.python-versionに基づいて自動で切り替わります。
 
 ```bash
-yarn run setup:cert-issuer
-
-or
-
 git clone https://github.com/blockchain-certificates/cert-issuer.git -b v3.3.0 .tmp/cert-issuer 
 cd .tmp/cert-issuer
 python setup.py experimental --blockchain=ethereum install
